I was doing a little driver compare and hacking to see if any of these other drivers had fixes that would improve this project.
In the end I realized the driver really needs to check the enc28j60 rev and take action for the errata for the given chip.... since there are multiple chips and it's a shared driver.
Also, the errata changes should allow a project to compile for there chip only.

I spent about a two days time on this over a few weeks...